requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cssselect==1.2.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.4
jieba3k==0.35.1
joblib==1.3.2
lxml==4.9.3
mediapipe==0.8.11
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.0
pandas==2.1.0
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
